Include timestamps in Redis responses
Timestamps are useful for calculating how long a prediction has taken to
run. There are other ways we could get this data (using HTTP streams
instead of queueing, and doing accurate timing frm the outside), but
this is probably the simplest to get started with and arguably the
simplest for all sorts of users in the future.

Signed-off-by: Dominic Baggott <dominic.baggott@gmail.com>